refactor validators to accept chain, following immutability
