+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/validate/CertificateChainValidator.kt
@@ -1,25 +1,23 @@
 package at.asitplus.signum.supreme.validate
 
-import at.asitplus.catchingUnwrapped
 import at.asitplus.signum.ExperimentalPkiApi
-import at.asitplus.signum.HazardousMaterials
 import at.asitplus.signum.KeyUsageException
-import at.asitplus.signum.indispensable.asn1.Asn1Exception
-import at.asitplus.signum.indispensable.asn1.KnownOIDs
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
-import at.asitplus.signum.indispensable.asn1.anyPolicy
 import at.asitplus.signum.indispensable.pki.CertificateChain
 import at.asitplus.signum.indispensable.pki.X509Certificate
-import at.asitplus.signum.indispensable.pki.leaf
 import at.asitplus.signum.indispensable.pki.pkiExtensions.BasicConstraintsExtension
 import at.asitplus.signum.indispensable.pki.pkiExtensions.KeyUsage
 import at.asitplus.signum.indispensable.pki.pkiExtensions.KeyUsageExtension
-import at.asitplus.signum.indispensable.pki.root
-import at.asitplus.signum.indispensable.pki.validate.*
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.contract
 import kotlin.time.Clock
 import kotlin.time.Instant
+
+interface CertificateChainValidator {
+    suspend fun validate(
+        chain: CertificateChain,
+        context: CertificateValidationContext,
+        checkedCriticalExtensions: MutableMap<X509Certificate, MutableSet<ObjectIdentifier>>
+    )
+}
 
 @OptIn(ExperimentalPkiApi::class)
 class CertificateValidationContext(
@@ -48,238 +46,3 @@ class CertificateValidationContext(
     },
     val supportRevocationChecking: Boolean = false
 )
-
-/**
- * Represents the result of validating a certificate chain.
- */
-sealed class CertificateValidationResult {
-
-    /** Indicates whether the certificate chain is fully valid (no validator failures). */
-    @OptIn(ExperimentalContracts::class)
-    val isValid: Boolean
-        get() {
-            contract {
-                returns(true) implies (this@CertificateValidationResult is Success)
-                returns(false) implies (this@CertificateValidationResult is Failure)
-            }
-            return this is Success
-        }
-
-    companion object {
-        internal operator fun invoke(
-            rootPolicyNode: PolicyNode?,
-            subject: X509Certificate,
-            validatorFailures: List<ValidatorFailure>
-        ) = if (validatorFailures.isNotEmpty()) {
-            Failure(validatorFailures)
-        } else Success(rootPolicyNode, subject)
-    }
-
-    data class Success(val rootPolicyNode: PolicyNode?, val subject: X509Certificate) : CertificateValidationResult()
-    data class Failure(val validatorFailures: List<ValidatorFailure>) : CertificateValidationResult() {
-        init {
-            require(validatorFailures.isNotEmpty()) { "Certificate validation failure must not be empty" }
-        }
-    }
-}
-
-/**
- * Represents a failure encountered by a specific validator during certificate chain validation.
- */
-data class ValidatorFailure(
-    val validatorName: String,
-    val validator: CertificateValidator? = null,
-    val errorMessage: String,
-    val certificateIndex: Int,
-    val cause: Throwable? = null
-)
-
-/**
- * Represents a factory for creating certificate validators from a given certificate chain
- * and a validation context. A factory is needed because validation is inherently stateful and individual
- * validators carry mutable state.
- *
- * @param generate A function that takes a `CertificateChain` and a `CertificateValidationContext`,
- *                 and returns a list of `CertificateValidator` instances to be applied.
- */
-fun interface ValidatorFactory {
-    fun CertificateChain.generate(
-        context: CertificateValidationContext
-    ): List<CertificateValidator>
-
-    companion object {
-        val RFC5280: ValidatorFactory =
-            ValidatorFactory { context -> defineRFC5280Validators(context, this) }
-    }
-}
-
-
-/**
- * Performs a custom, potentially unsafe validation of this [CertificateChain] using the provided [validatorFactory].
- *
- * This function executes all validators generated by the [validatorFactory] against every certificate in the chain.
- * It does not automatically enforce RFC 5280 compliance — any custom validators can be provided
- *
- * @return a [CertificateValidationResult] containing the resulting policy tree and
- * the end-entity certificate, or a list of [ValidatorFailure]s for a chain that failed to validate.
- *
- * @throws Asn1Exception on structurally invalid input
- */
-@OptIn(HazardousMaterials::class)
-@ExperimentalPkiApi
-@Throws(Asn1Exception::class, kotlin.coroutines.cancellation.CancellationException::class)
-suspend fun CertificateChain.validate(
-    validatorFactory: ValidatorFactory,
-    context: CertificateValidationContext = CertificateValidationContext()
-): CertificateValidationResult {
-
-    val validators = with(validatorFactory) { this@validate.generate(context) }
-    val processingChain = if (context.allowIncludedTrustAnchor && context.trustAnchors.any {
-            it.matchesCertificate(this.root)
-        }) this.dropLast(1) else this
-
-    val activeValidators = validators.toMutableList()
-    val validatorFailures = mutableListOf<ValidatorFailure>()
-    val trustAnchorValidator = activeValidators.filterIsInstance<TrustAnchorValidator>().firstOrNull()
-    val nameConstraintsValidator = activeValidators.filterIsInstance<NameConstraintsValidator>().firstOrNull()
-
-    trustAnchorValidator?.let { trustAnchorValidator ->
-        catchingUnwrapped {
-            processingChain.forEach {
-                trustAnchorValidator.check(it, it.criticalExtensionOids.toMutableSet())
-            }
-        }.onFailure {
-            validatorFailures.add(
-                ValidatorFailure(
-                    TrustAnchorValidator::class.simpleName!!,
-                    trustAnchorValidator,
-                    it.message ?: "Trust Anchor validation failed.",
-                    -1,
-                    it
-                )
-            )
-        }
-
-        nameConstraintsValidator?.previousNameConstraints = trustAnchorValidator.trustAnchor?.nameConstraints
-        activeValidators.remove(trustAnchorValidator)
-    }
-
-    processingChain.reversed().forEachIndexed { i, cert ->
-        val checkedCriticalExtensions = mutableSetOf<ObjectIdentifier>()
-
-        val validatorIterator = activeValidators.iterator()
-        for (currValidator in validatorIterator) {
-            try {
-                currValidator.check(cert, checkedCriticalExtensions)
-            } catch (e: Throwable) {
-                validatorIterator.remove()
-                validatorFailures.add(
-                    ValidatorFailure(
-                        currValidator::class.simpleName!!,
-                        currValidator,
-                        e.message ?: "Validation failed.",
-                        i + 1,
-                        e
-                    )
-                )
-            }
-        }
-        verifyCriticalExtensions(checkedCriticalExtensions, cert, i, validatorFailures)
-    }
-    return CertificateValidationResult((validators.find { it is PolicyValidator } as? PolicyValidator)?.rootNode,
-        this.leaf,
-        validatorFailures)
-}
-
-/**
- * Performs an RFC 5280-compliant validation of this [CertificateChain] using default validators.
- *
- * This function automatically adds all mandatory validators required for standard X.509
- * path validation according to RFC 5280, including:
- * [PolicyValidator] – for certificate policy processing
- * [NameConstraintsValidator] – for name constraint enforcement
- * [KeyUsageValidator] – for key usage and extended key usage checks
- * [BasicConstraintsValidator] – for CA and path length validation
- * [ChainValidator] – for signature chain integrity
- * [TimeValidityValidator] – for certificate validity period
- * [CertValidityValidator] – for checking whether the certificate is constructed correctly, since some components are decoded too leniently
- *
- * [TrustAnchorValidator] is also executed to ensure that the chain terminates
- * in a trusted root or intermediate authority.
- * [KeyIdentifierValidator] is added to validate Subject and Authority key identifiers
- *
- * Validation results are collected rather than throwing exceptions
- *
- * @return a [CertificateValidationResult] containing the resulting policy tree and
- * the end-entity certificate, or a list of [ValidatorFailure]s for a chain that failed to validate.
- *
- * @throws Asn1Exception on structurally invalid input
- */
-
-@Throws(Asn1Exception::class, kotlin.coroutines.cancellation.CancellationException::class)
-@ExperimentalPkiApi
-suspend fun CertificateChain.validate(
-    context: CertificateValidationContext = CertificateValidationContext()
-): CertificateValidationResult = validate(ValidatorFactory.RFC5280, context)
-
-/** Constructs list of default validator used in RFC5280 validation based on [context] */
-private fun defineRFC5280Validators(
-    context: CertificateValidationContext,
-    chain: CertificateChain
-): List<CertificateValidator> {
-    val (pathLen, processingChain) =
-        if (context.allowIncludedTrustAnchor && context.trustAnchors.any { it.matchesCertificate(chain.root) }) {
-            chain.size - 1 to chain.dropLast(1)
-        } else {
-            chain.size to chain
-        }
-
-    return listOf(
-        PolicyValidator(
-            initialPolicies = context.initialPolicies,
-            expPolicyRequired = context.explicitPolicyRequired,
-            polMappingInhibited = context.policyMappingInhibited,
-            anyPolicyInhibited = context.anyPolicyInhibited,
-            certPathLen = pathLen,
-            rejectPolicyQualifiers = context.policyQualifiersRejected,
-            rootNode = PolicyNode(
-                parent = null,
-                validPolicy = KnownOIDs.anyPolicy,
-                criticalityIndicator = false,
-                expectedPolicySet = setOf(KnownOIDs.anyPolicy),
-                generatedByPolicyMapping = false
-            )
-        ),
-        CertValidityValidator(),
-        NameConstraintsValidator(pathLen),
-        KeyUsageValidator(pathLen, expectedEku = context.expectedEku, leafKeyUsageCheck = context.leafKeyUsageCheck, supportRevocationChecking = context.supportRevocationChecking),
-        BasicConstraintsValidator(pathLen),
-        ChainValidator(processingChain.reversed()),
-        TimeValidityValidator(context.date, certChain = processingChain.reversed()),
-        TrustAnchorValidator(context.trustAnchors, processingChain, date = context.date),
-        KeyIdentifierValidator()
-    )
-}
-
-/**
- * Checks if there are any unhandled critical extensions remaining,
- * which would indicate that the current validators do not support them.
- */
-private fun verifyCriticalExtensions(
-    checkedCriticalExtensions: Set<ObjectIdentifier>,
-    cert: X509Certificate,
-    certificateIndex: Int,
-    failures: MutableList<ValidatorFailure>
-) {
-    val unsupportedCriticalExtensions = cert.criticalExtensionOids - checkedCriticalExtensions
-    if (unsupportedCriticalExtensions.isNotEmpty() && failures.none { it.validatorName == "CriticalCertificateExtensions" }) {
-        failures.add(
-            ValidatorFailure(
-                validatorName = "CriticalCertificateExtensions",
-                errorMessage = "Unsupported critical extensions: $unsupportedCriticalExtensions",
-                certificateIndex = certificateIndex
-            )
-        )
-    }
-}
-
