+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/validate/CertificateValidator.kt
@@ -1,12 +1,13 @@
-package at.asitplus.signum.indispensable.pki.validate
+package at.asitplus.signum.supreme.validate
 
 import at.asitplus.signum.CertificateException
 import at.asitplus.signum.ExperimentalPkiApi
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
+import at.asitplus.signum.indispensable.pki.CertificateChain
 import at.asitplus.signum.indispensable.pki.X509Certificate
 import kotlin.coroutines.cancellation.CancellationException
 
-interface CertificateValidator {
+interface CertificateValidator : CertificateChainValidator {
 
     /**
      * Performs certificate validation for the given certificate
@@ -18,5 +19,18 @@ interface CertificateValidator {
      */
     @ExperimentalPkiApi
     @Throws(Throwable::class)
-    suspend fun check(currCert: X509Certificate, checkedCriticalExtensions: MutableSet<ObjectIdentifier>)
+    suspend fun check(currCert: X509Certificate) : Set<ObjectIdentifier>
+
+    @OptIn(ExperimentalPkiApi::class)
+    override suspend fun validate(
+        chain: CertificateChain,
+        context: CertificateValidationContext,
+        checkedCriticalExtensions: MutableMap<X509Certificate, MutableSet<ObjectIdentifier>>
+    ) {
+        chain.forEach {
+            checkedCriticalExtensions
+                .getOrPut(it) { mutableSetOf() }
+                .addAll(check(it))
+        }
+    }
 }
