+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/validate/PolicyNode.kt
@@ -1,4 +1,4 @@
-package at.asitplus.signum.indispensable.pki.validate
+package at.asitplus.signum.supreme.validate
 
 import at.asitplus.signum.ExperimentalPkiApi
 import at.asitplus.signum.indispensable.asn1.*
