+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/validate/PolicyValidator.kt
@@ -1,10 +1,13 @@
-package at.asitplus.signum.indispensable.pki.validate
+package at.asitplus.signum.supreme.validate
 
 import at.asitplus.signum.CertificatePolicyException
 import at.asitplus.signum.ExperimentalPkiApi
 import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
+import at.asitplus.signum.indispensable.asn1.inhibitAnyPolicy
+import at.asitplus.signum.indispensable.asn1.policyMapping
 import at.asitplus.signum.indispensable.asn1.toBigInteger
+import at.asitplus.signum.indispensable.pki.CertificateChain
 import at.asitplus.signum.indispensable.pki.X509Certificate
 import at.asitplus.signum.indispensable.pki.pkiExtensions.CertificatePoliciesExtension
 import at.asitplus.signum.indispensable.pki.pkiExtensions.InhibitAnyPolicyExtension
@@ -16,21 +19,8 @@ import at.asitplus.signum.indispensable.pki.pkiExtensions.PolicyQualifierInfo
 * PolicyValidator checks policy information on X509Certificate path
 */
 class PolicyValidator(
-    initialPolicies: Set<ObjectIdentifier>,
-    expPolicyRequired: Boolean,
-    polMappingInhibited: Boolean,
-    anyPolicyInhibited: Boolean,
-    private val certPathLen: Int,
-    private val rejectPolicyQualifiers: Boolean,
     var rootNode: PolicyNode?
-) : CertificateValidator {
-    private val initPolicies: Set<ObjectIdentifier> =
-        initialPolicies.ifEmpty { setOf(KnownOIDs.anyPolicy) }.toSet()
-    private var explicitPolicy: Int = 0
-    private var policyMapping: Int = 0
-    private var inhibitAnyPolicy: Int = 0
-    private var certIndex: Int = 0
-
+) : CertificateChainValidator {
     private var supportedExtensions: Set<ObjectIdentifier> = setOf(
         KnownOIDs.certificatePolicies_2_5_29_32,
         KnownOIDs.policyMappings,
@@ -38,35 +28,45 @@ class PolicyValidator(
         KnownOIDs.inhibitAnyPolicy
     )
 
-    init {
-        certIndex = 1
-        explicitPolicy = if (expPolicyRequired) 0 else certPathLen + 1
-        policyMapping = if (polMappingInhibited) 0 else certPathLen + 1
-        inhibitAnyPolicy = if (anyPolicyInhibited) 0 else certPathLen + 1
-    }
+    var explicitPolicy = 0
 
-    @ExperimentalPkiApi
-    override suspend fun check(currCert: X509Certificate, checkedCriticalExtensions: MutableSet<ObjectIdentifier>) {
-        checkedCriticalExtensions.addAll(supportedExtensions)
+    override suspend fun validate(
+        chain: CertificateChain,
+        context: CertificateValidationContext,
+        checkedCriticalExtensions: MutableMap<X509Certificate, MutableSet<ObjectIdentifier>>
+    ) {
+        var certIndex = 1
+        val certPathLen = chain.size
+        val rejectPolicyQualifiers = context.policyQualifiersRejected
+        val initPolicies: Set<ObjectIdentifier> =
+            context.initialPolicies.ifEmpty { setOf(KnownOIDs.anyPolicy) }.toSet()
+        explicitPolicy = if (context.explicitPolicyRequired) 0 else certPathLen + 1
+        var policyMapping = if (context.policyMappingInhibited) 0 else certPathLen + 1
+        var inhibitAnyPolicy = if (context.anyPolicyInhibited) 0 else certPathLen + 1
 
-        rootNode = processPolicies(
-            certIndex,
-            initPolicies,
-            explicitPolicy,
-            policyMapping,
-            inhibitAnyPolicy,
-            rejectPolicyQualifiers,
-            rootNode,
-            currCert,
-            certIndex == certPathLen
-        )
+        for (currCert in chain) {
+            checkedCriticalExtensions
+                .getOrPut(currCert) { mutableSetOf() }
+                .addAll(supportedExtensions)
+            rootNode = processPolicies(
+                certIndex,
+                initPolicies,
+                explicitPolicy,
+                policyMapping,
+                inhibitAnyPolicy,
+                rejectPolicyQualifiers,
+                rootNode,
+                currCert,
+                certIndex == certPathLen
+            )
 
-        if (certIndex != certPathLen) {
-            explicitPolicy = updateExplicitPolicy(explicitPolicy, currCert, false)
-            policyMapping = updatePolicyMapping(policyMapping, currCert)
-            inhibitAnyPolicy = updateInhibitAnyPolicy(inhibitAnyPolicy, currCert)
+            if (certIndex != certPathLen) {
+                explicitPolicy = updateExplicitPolicy(explicitPolicy, currCert, false)
+                policyMapping = updatePolicyMapping(policyMapping, currCert)
+                inhibitAnyPolicy = updateInhibitAnyPolicy(inhibitAnyPolicy, currCert)
+            }
+            certIndex++
         }
-        certIndex++
     }
 
     /**
@@ -239,7 +239,7 @@ class PolicyValidator(
 
         if (root != null && !initialPolicies.contains(KnownOIDs.anyPolicy)) {
             root = policyExtension?.let {
-                removeInvalidNodes(root!!, certIndex, initialPolicies, it)
+                removeInvalidNodes(root, certIndex, initialPolicies, it)
             }
 
             // RFC 5280: 6.1.5 (g)(iii)
