+++ supreme/src/jvmTest/kotlin/at/asitplus/signum/supreme/validate/ValidationApiTest.kt
@@ -1,155 +1,151 @@
-package at.asitplus.signum.supreme.validate
-
-import at.asitplus.signum.ExperimentalPkiApi
-import at.asitplus.signum.InvalidCertificateValidityPeriodException
-import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
-import at.asitplus.signum.indispensable.pki.CertificateChain
-import at.asitplus.signum.indispensable.pki.X509Certificate
-import at.asitplus.signum.indispensable.pki.leaf
-import at.asitplus.signum.indispensable.pki.root
-import at.asitplus.signum.indispensable.pki.validate.CertValidityValidator
-import at.asitplus.signum.indispensable.pki.validate.CertificateValidator
-import at.asitplus.signum.indispensable.pki.validate.KeyIdentifierValidator
-import at.asitplus.signum.indispensable.pki.validate.TimeValidityValidator
-import at.asitplus.testballoon.invoke
-import de.infix.testBalloon.framework.core.testSuite
-import io.kotest.matchers.shouldBe
-import kotlin.time.Clock
-import kotlin.time.Instant
-
-
-@OptIn(ExperimentalPkiApi::class)
-val ValidationApiTest by testSuite{
-
-    val testSuiteLimbo = json.decodeFromString<LimboSuite>(resourceText("limbo.json"))
-
-    // we are sending trust anchors as part of the chain, but expecting that they are included in system trust store, so they will be omitted from the chain during validation
-    context("online testcases") {
-        val onlineTests = testSuiteLimbo.testcases.filter {
-            it.id.contains("online", ignoreCase = true)
-                    && !it.id.contains("online::stackoverflow.com", ignoreCase = true)
-
-        }
-        onlineTests.forEach {
-            test("Online testcase validated using system trust store: ${it.id}") {
-                val trustAnchors = it.trusted_certs.map { pem ->
-                    X509Certificate.decodeFromPem(pem).getOrThrow()
-                }
-
-                val intermediates = it.untrusted_intermediates.map { pem ->
-                    X509Certificate.decodeFromPem(pem).getOrThrow()
-                }
-
-                val leaf = X509Certificate.decodeFromPem(it.peer_certificate).getOrThrow()
-
-                val chain: CertificateChain = listOf(leaf) + intermediates.reversed() + trustAnchors.reversed()
-                val validationTime = it.validation_time?.let(Instant::parse) ?: Clock.System.now()
-
-                val context = CertificateValidationContext(
-                    allowIncludedTrustAnchor = true, // default is true, but for the clarity
-                    expectedEku = it.extended_key_usage.mapNotNull { extendedKeyUsages[it] }.toSet(),
-                    date = validationTime
-                )
-
-                val result = chain.validate(context)
-
-                if (it.expected_result == "FAILURE") {
-                    result.isValid shouldBe false
-                } else {
-                    result.isValid shouldBe true
-                }
-
-            }
-        }
-    }
-
-    "Validate Attestation Proof With RFC5280 And Custom Rules" {
-        val attestationProofB64 = listOf(
-            """MIIC+jCCAqCgAwIBAgIBATAKBggqhkjOPQQDAjCBiDELMAkGA1UEBhMCVVMxEzARBgNVB
-                    AgMCkNhbGlmb3JuaWExFTATBgNVBAoMDEdvb2dsZSwgSW5jLjEQMA4GA1UECwwHQW5kcm9pZ
-                    DE7MDkGA1UEAwwyQW5kcm9pZCBLZXlzdG9yZSBTb2Z0d2FyZSBBdHRlc3RhdGlvbiBJbnRlc
-                    m1lZGlhdGUwHhcNNzAwMTAxMDAwMDAwWhcNNjkxMjMxMjM1OTU5WjAfMR0wGwYDVQQDDBRBb
-                    mRyb2lkIEtleXN0b3JlIEtleTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABIEAthaOZ2+nZ
-                    ZyYdoeLYNL5yZozzfMdrfrZcG2RI1juriVparubkzxZGCs4KcReh1aDhWFsDxQWYAeJLcFN8
-                    rOjggFhMIIBXTALBgNVHQ8EBAMCB4AwggErBgorBgEEAdZ5AgERBIIBGzCCARcCAQQKAQACA
-                    SkKAQAEEETfQo1OyOc6bwoew974v2gEADCB8qEIMQYCAQICAQOiAwIBA6MEAgIBAKUIMQYCA
-                    QICAQSqAwIBAb+DdwIFAL+FPQgCBgGHj7zJmL+FPgMCAQC/hUBMMEoEIAAAAAAAAAAAAAAAA
-                    AAAAAAAAAAAAAAAAAAAAAAAAAAAAQEACgECBCAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-                    AAAAAAAAL+FQQUCAwGtsL+FQgUCAwMVG7+FRU8ETTBLMSUwIwQeYXQuYXNpdHBsdXMuYXR0Z
-                    XN0YXRpb25fY2xpZW50AgEBMSIEIDS5dixNbJDUhDGUDFe95zFCWLJkIO/hasf3J08NMwrVM
-                    AAwHwYDVR0jBBgwFoAUP/ys1hqxOp6BILjVJRzFZbsekakwCgYIKoZIzj0EAwIDSAAwRQIgW
-                    CsSigJsOLe9hli462AL/TuPqLuuIKelSVEe/PsnrWUCIQC+JExSC5l3slEBhDKxMD3otjwr0
-                    DK0Jav50CzyK80ILg==
-                    """,
-            """MIICeDCCAh6gAwIBAgICEAEwCgYIKoZIzj0EAwIwgZgxCzAJBgNVBAYTAlVTMRMwEQYDV
-                    QQIDApDYWxpZm9ybmlhMRYwFAYDVQQHDA1Nb3VudGFpbiBWaWV3MRUwEwYDVQQKDAxHb29nb
-                    GUsIEluYy4xEDAOBgNVBAsMB0FuZHJvaWQxMzAxBgNVBAMMKkFuZHJvaWQgS2V5c3RvcmUgU
-                    29mdHdhcmUgQXR0ZXN0YXRpb24gUm9vdDAeFw0xNjAxMTEwMDQ2MDlaFw0yNjAxMDgwMDQ2M
-                    DlaMIGIMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTEVMBMGA1UECgwMR29vZ
-                    2xlLCBJbmMuMRAwDgYDVQQLDAdBbmRyb2lkMTswOQYDVQQDDDJBbmRyb2lkIEtleXN0b3JlI
-                    FNvZnR3YXJlIEF0dGVzdGF0aW9uIEludGVybWVkaWF0ZTBZMBMGByqGSM49AgEGCCqGSM49A
-                    wEHA0IABOueefhCY1msyyqRTImGzHCtkGaTgqlzJhP+rMv4ISdMIXSXSir+pblNf2bU4GUQZ
-                    jW8U7ego6ZxWD7bPhGuEBSjZjBkMB0GA1UdDgQWBBQ//KzWGrE6noEguNUlHMVlux6RqTAfB
-                    gNVHSMEGDAWgBTIrel3TEXDo88NFhDkeUM6IVowzzASBgNVHRMBAf8ECDAGAQH/AgEAMA4GA
-                    1UdDwEB/wQEAwIChDAKBggqhkjOPQQDAgNIADBFAiBLipt77oK8wDOHri/AiZi03cONqycqR
-                    Z9pDMfDktQPjgIhAO7aAV229DLp1IQ7YkyUBO86fMy9Xvsiu+f+uXc/WT/7
-                    """,
-            """MIICizCCAjKgAwIBAgIJAKIFntEOQ1tXMAoGCCqGSM49BAMCMIGYMQswCQYDVQQGEwJVU
-                    zETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwNTW91bnRhaW4gVmlldzEVMBMGA1UEC
-                    gwMR29vZ2xlLCBJbmMuMRAwDgYDVQQLDAdBbmRyb2lkMTMwMQYDVQQDDCpBbmRyb2lkIEtle
-                    XN0b3JlIFNvZnR3YXJlIEF0dGVzdGF0aW9uIFJvb3QwHhcNMTYwMTExMDA0MzUwWhcNMzYwM
-                    TA2MDA0MzUwWjCBmDELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVB
-                    AcMDU1vdW50YWluIFZpZXcxFTATBgNVBAoMDEdvb2dsZSwgSW5jLjEQMA4GA1UECwwHQW5kc
-                    m9pZDEzMDEGA1UEAwwqQW5kcm9pZCBLZXlzdG9yZSBTb2Z0d2FyZSBBdHRlc3RhdGlvbiBSb
-                    290MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE7l1ex+HA220Dpn7mthvsTWpdamguD/9/S
-                    Q59dx9EIm29sa/6FsvHrcV30lacqrewLVQBXT5DKyqO107sSHVBpKNjMGEwHQYDVR0OBBYEF
-                    Mit6XdMRcOjzw0WEOR5QzohWjDPMB8GA1UdIwQYMBaAFMit6XdMRcOjzw0WEOR5QzohWjDPM
-                    A8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgKEMAoGCCqGSM49BAMCA0cAMEQCIDUho
-                    ++LNEYenNVg8x1YiSBq3KNlQfYNns6KGYxmSGB7AiBNC/NR2TB8fVvaNTQdqEcbY6WFZTytT
-                    ySn502vQX3xvw==
-                    """
-        )
-
-        val chain: CertificateChain = attestationProofB64.map {
-            X509Certificate.decodeFromByteArray(java.util.Base64.getMimeDecoder()
-                .decode(it))!!
-        }
-
-        chain.validate(
-            CertificateValidationContext(
-                trustAnchors = setOf(TrustAnchor.Certificate(chain.root)),
-            )
-        ).isValid shouldBe false
-
-        val customValidatorFactory = ValidatorFactory { context ->
-            val validators = ValidatorFactory.RFC5280.run { chain.generate(context) }.toMutableList()
-            validators.removeAll { it is CertValidityValidator || it is KeyIdentifierValidator || it is TimeValidityValidator }
-            validators.add(AttestationTimeValidator(this))
-            validators
-        }
-
-        chain.validate(
-            customValidatorFactory,
-            CertificateValidationContext(
-                trustAnchors = setOf(TrustAnchor.Certificate(chain.root))
-            )
-        ).isValid shouldBe true
-    }
-}
-
-// Example of custom validator, it checks validity only in intermediate certificate
-class AttestationTimeValidator(
-    val certChain: CertificateChain
-) : CertificateValidator {
-
-    @ExperimentalPkiApi
-    override suspend fun check(
-        currCert: X509Certificate,
-        checkedCriticalExtensions: MutableSet<ObjectIdentifier>
-    ) {
-        if (currCert != certChain.leaf) {
-            if (!currCert.isValidAt(currCert.tbsCertificate.validUntil.instant)) throw InvalidCertificateValidityPeriodException("Certificate is not valid")
-        }
-    }
-
-}
+//package at.asitplus.signum.supreme.validate
+//
+//import at.asitplus.signum.ExperimentalPkiApi
+//import at.asitplus.signum.InvalidCertificateValidityPeriodException
+//import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
+//import at.asitplus.signum.indispensable.pki.CertificateChain
+//import at.asitplus.signum.indispensable.pki.X509Certificate
+//import at.asitplus.signum.indispensable.pki.leaf
+//import at.asitplus.signum.indispensable.pki.root
+//import at.asitplus.testballoon.invoke
+//import de.infix.testBalloon.framework.core.testSuite
+//import io.kotest.matchers.shouldBe
+//import kotlin.time.Clock
+//import kotlin.time.Instant
+//
+//
+//@OptIn(ExperimentalPkiApi::class)
+//val ValidationApiTest by testSuite{
+//
+//    val testSuiteLimbo = json.decodeFromString<LimboSuite>(resourceText("limbo.json"))
+//
+//    // we are sending trust anchors as part of the chain, but expecting that they are included in system trust store, so they will be omitted from the chain during validation
+//    context("online testcases") {
+//        val onlineTests = testSuiteLimbo.testcases.filter {
+//            it.id.contains("online", ignoreCase = true)
+//                    && !it.id.contains("online::stackoverflow.com", ignoreCase = true)
+//
+//        }
+//        onlineTests.forEach {
+//            test("Online testcase validated using system trust store: ${it.id}") {
+//                val trustAnchors = it.trusted_certs.map { pem ->
+//                    X509Certificate.decodeFromPem(pem).getOrThrow()
+//                }
+//
+//                val intermediates = it.untrusted_intermediates.map { pem ->
+//                    X509Certificate.decodeFromPem(pem).getOrThrow()
+//                }
+//
+//                val leaf = X509Certificate.decodeFromPem(it.peer_certificate).getOrThrow()
+//
+//                val chain: CertificateChain = listOf(leaf) + intermediates.reversed() + trustAnchors.reversed()
+//                val validationTime = it.validation_time?.let(Instant::parse) ?: Clock.System.now()
+//
+//                val context = CertificateValidationContext(
+//                    allowIncludedTrustAnchor = true, // default is true, but for the clarity
+//                    expectedEku = it.extended_key_usage.mapNotNull { extendedKeyUsages[it] }.toSet(),
+//                    date = validationTime
+//                )
+//
+//                val result = chain.validate(context)
+//
+//                if (it.expected_result == "FAILURE") {
+//                    result.isValid shouldBe false
+//                } else {
+//                    result.isValid shouldBe true
+//                }
+//
+//            }
+//        }
+//    }
+//
+//    "Validate Attestation Proof With RFC5280 And Custom Rules" {
+//        val attestationProofB64 = listOf(
+//            """MIIC+jCCAqCgAwIBAgIBATAKBggqhkjOPQQDAjCBiDELMAkGA1UEBhMCVVMxEzARBgNVB
+//                    AgMCkNhbGlmb3JuaWExFTATBgNVBAoMDEdvb2dsZSwgSW5jLjEQMA4GA1UECwwHQW5kcm9pZ
+//                    DE7MDkGA1UEAwwyQW5kcm9pZCBLZXlzdG9yZSBTb2Z0d2FyZSBBdHRlc3RhdGlvbiBJbnRlc
+//                    m1lZGlhdGUwHhcNNzAwMTAxMDAwMDAwWhcNNjkxMjMxMjM1OTU5WjAfMR0wGwYDVQQDDBRBb
+//                    mRyb2lkIEtleXN0b3JlIEtleTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABIEAthaOZ2+nZ
+//                    ZyYdoeLYNL5yZozzfMdrfrZcG2RI1juriVparubkzxZGCs4KcReh1aDhWFsDxQWYAeJLcFN8
+//                    rOjggFhMIIBXTALBgNVHQ8EBAMCB4AwggErBgorBgEEAdZ5AgERBIIBGzCCARcCAQQKAQACA
+//                    SkKAQAEEETfQo1OyOc6bwoew974v2gEADCB8qEIMQYCAQICAQOiAwIBA6MEAgIBAKUIMQYCA
+//                    QICAQSqAwIBAb+DdwIFAL+FPQgCBgGHj7zJmL+FPgMCAQC/hUBMMEoEIAAAAAAAAAAAAAAAA
+//                    AAAAAAAAAAAAAAAAAAAAAAAAAAAAQEACgECBCAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+//                    AAAAAAAAL+FQQUCAwGtsL+FQgUCAwMVG7+FRU8ETTBLMSUwIwQeYXQuYXNpdHBsdXMuYXR0Z
+//                    XN0YXRpb25fY2xpZW50AgEBMSIEIDS5dixNbJDUhDGUDFe95zFCWLJkIO/hasf3J08NMwrVM
+//                    AAwHwYDVR0jBBgwFoAUP/ys1hqxOp6BILjVJRzFZbsekakwCgYIKoZIzj0EAwIDSAAwRQIgW
+//                    CsSigJsOLe9hli462AL/TuPqLuuIKelSVEe/PsnrWUCIQC+JExSC5l3slEBhDKxMD3otjwr0
+//                    DK0Jav50CzyK80ILg==
+//                    """,
+//            """MIICeDCCAh6gAwIBAgICEAEwCgYIKoZIzj0EAwIwgZgxCzAJBgNVBAYTAlVTMRMwEQYDV
+//                    QQIDApDYWxpZm9ybmlhMRYwFAYDVQQHDA1Nb3VudGFpbiBWaWV3MRUwEwYDVQQKDAxHb29nb
+//                    GUsIEluYy4xEDAOBgNVBAsMB0FuZHJvaWQxMzAxBgNVBAMMKkFuZHJvaWQgS2V5c3RvcmUgU
+//                    29mdHdhcmUgQXR0ZXN0YXRpb24gUm9vdDAeFw0xNjAxMTEwMDQ2MDlaFw0yNjAxMDgwMDQ2M
+//                    DlaMIGIMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTEVMBMGA1UECgwMR29vZ
+//                    2xlLCBJbmMuMRAwDgYDVQQLDAdBbmRyb2lkMTswOQYDVQQDDDJBbmRyb2lkIEtleXN0b3JlI
+//                    FNvZnR3YXJlIEF0dGVzdGF0aW9uIEludGVybWVkaWF0ZTBZMBMGByqGSM49AgEGCCqGSM49A
+//                    wEHA0IABOueefhCY1msyyqRTImGzHCtkGaTgqlzJhP+rMv4ISdMIXSXSir+pblNf2bU4GUQZ
+//                    jW8U7ego6ZxWD7bPhGuEBSjZjBkMB0GA1UdDgQWBBQ//KzWGrE6noEguNUlHMVlux6RqTAfB
+//                    gNVHSMEGDAWgBTIrel3TEXDo88NFhDkeUM6IVowzzASBgNVHRMBAf8ECDAGAQH/AgEAMA4GA
+//                    1UdDwEB/wQEAwIChDAKBggqhkjOPQQDAgNIADBFAiBLipt77oK8wDOHri/AiZi03cONqycqR
+//                    Z9pDMfDktQPjgIhAO7aAV229DLp1IQ7YkyUBO86fMy9Xvsiu+f+uXc/WT/7
+//                    """,
+//            """MIICizCCAjKgAwIBAgIJAKIFntEOQ1tXMAoGCCqGSM49BAMCMIGYMQswCQYDVQQGEwJVU
+//                    zETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwNTW91bnRhaW4gVmlldzEVMBMGA1UEC
+//                    gwMR29vZ2xlLCBJbmMuMRAwDgYDVQQLDAdBbmRyb2lkMTMwMQYDVQQDDCpBbmRyb2lkIEtle
+//                    XN0b3JlIFNvZnR3YXJlIEF0dGVzdGF0aW9uIFJvb3QwHhcNMTYwMTExMDA0MzUwWhcNMzYwM
+//                    TA2MDA0MzUwWjCBmDELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVB
+//                    AcMDU1vdW50YWluIFZpZXcxFTATBgNVBAoMDEdvb2dsZSwgSW5jLjEQMA4GA1UECwwHQW5kc
+//                    m9pZDEzMDEGA1UEAwwqQW5kcm9pZCBLZXlzdG9yZSBTb2Z0d2FyZSBBdHRlc3RhdGlvbiBSb
+//                    290MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE7l1ex+HA220Dpn7mthvsTWpdamguD/9/S
+//                    Q59dx9EIm29sa/6FsvHrcV30lacqrewLVQBXT5DKyqO107sSHVBpKNjMGEwHQYDVR0OBBYEF
+//                    Mit6XdMRcOjzw0WEOR5QzohWjDPMB8GA1UdIwQYMBaAFMit6XdMRcOjzw0WEOR5QzohWjDPM
+//                    A8GA1UdEwEB/wQFMAMBAf8wDgYDVR0PAQH/BAQDAgKEMAoGCCqGSM49BAMCA0cAMEQCIDUho
+//                    ++LNEYenNVg8x1YiSBq3KNlQfYNns6KGYxmSGB7AiBNC/NR2TB8fVvaNTQdqEcbY6WFZTytT
+//                    ySn502vQX3xvw==
+//                    """
+//        )
+//
+//        val chain: CertificateChain = attestationProofB64.map {
+//            X509Certificate.decodeFromByteArray(java.util.Base64.getMimeDecoder()
+//                .decode(it))!!
+//        }
+//
+//        chain.validate(
+//            CertificateValidationContext(
+//                trustAnchors = setOf(TrustAnchor.Certificate(chain.root)),
+//            )
+//        ).isValid shouldBe false
+//
+//        val customValidatorFactory = ValidatorFactory { context ->
+//            val validators = ValidatorFactory.RFC5280.run { chain.generate(context) }.toMutableList()
+//            validators.removeAll { it is CertValidityValidator || it is KeyIdentifierValidator || it is TimeValidityValidator }
+//            validators.add(AttestationTimeValidator(this))
+//            validators
+//        }
+//
+//        chain.validate(
+//            customValidatorFactory,
+//            CertificateValidationContext(
+//                trustAnchors = setOf(TrustAnchor.Certificate(chain.root))
+//            )
+//        ).isValid shouldBe true
+//    }
+//}
+//
+//// Example of custom validator, it checks validity only in intermediate certificate
+//class AttestationTimeValidator(
+//    val certChain: CertificateChain
+//) : CertificateValidator {
+//
+//    @ExperimentalPkiApi
+//    override suspend fun check(
+//        currCert: X509Certificate,
+//        checkedCriticalExtensions: MutableSet<ObjectIdentifier>
+//    ) {
+//        if (currCert != certChain.leaf) {
+//            if (!currCert.isValidAt(currCert.tbsCertificate.validUntil.instant)) throw InvalidCertificateValidityPeriodException("Certificate is not valid")
+//        }
+//    }
+//
+//}
